+++ templates/Logentries/show_entries.php
@@ -13,16 +13,26 @@
     </div>
     <p></p>
     <div class="table-responsive">
+        <p>
+            <ul>
+                <li><?= $this->Html->link('Show all log entries', ['controller' => 'logentries', 'action' => 'index']) ?></li>
+                <li><?= $this->Html->link('Show only errors (Type >= 50)', ['controller' => 'logentries', 'action' => 'errors']) ?></li>
+            </ul>
+            Available log types:<br>
+            10 - Notification<br>
+            20 - Sent email (not implemented yet)<br>
+            30 - Automated problem fixing (course reminders, etc.)<br>
+            50 - Non-fatal error<br>
+            90 - Fatal error<br>
+        </p>
         <table>
             <thead>
                 <tr>
                     <th align="left" style="padding: 5px"><?= $this->Paginator->sort('id') ?></th>
                     <th align="left" style="padding: 5px"><?= $this->Paginator->sort('logentry_code_id', 'Type') ?></th>
-                    <th align="left" style="padding: 5px"><?= $this->Paginator->sort('user_id') ?></th>
                     <th align="left" style="padding: 5px"><?= $this->Paginator->sort('source_name', 'Application Source') ?></th>
                     <th align="left" style="padding: 5px"><?= $this->Paginator->sort('subject', 'Action') ?></th>
                     <th align="left" style="padding: 5px"><?= $this->Paginator->sort('description', 'Details') ?></th>
-                    <th align="left" style="padding: 5px"><?= $this->Paginator->sort('cleared') ?></th>
                     <th align="left" style="padding: 5px"><?= $this->Paginator->sort('created', ['label' => 'Date & Time']) ?></th>
                 </tr>
             </thead>
@@ -31,19 +41,13 @@
                     <tr>
                         <td align="left" style="padding: 5px"><?= $this->Number->format($logentry->id) ?></td>
                         <td align="left" style="padding: 5px"><?= $logentry->logentry_code->id . ' - ' . $logentry->logentry_code->name ?></td>
-                        <td align="left" style="padding: 5px"><?= $this->Html->link(
-                                                                    $logentry->user->id,
-                                                                    ['controller' => 'Users', 'action' => 'view', $logentry->user->id],
-                                                                    ['title' => $logentry->user->first_name . ' ' . $logentry->user->last_name]
-                                                                ) ?></td>
                         <td align="left" style="padding: 5px"><?= h($logentry->source_name) ?></td>
                         <td align="left" style="padding: 5px"><?= h($logentry->subject) ?></td>
                         <td align="left" style="padding: 5px"><?= $this->Text->autoParagraph($logentry->description) ?></td>
-                        <td align="left" style="padding: 5px"><?= ($logentry->cleared ? "Yes" : "No") ?></td>
                         <td align="left" style="padding: 5px"><?= h($logentry->created->i18nFormat('yyyy-MM-dd HH:mm')) ?> UTC</td>
                     </tr>
                     <tr>
-                        <td colspan="8">
+                        <td colspan="6">
                             <hr>
                         </td>
                     </tr>